auth functions integrated
