+++ client/src/components/Logout/Logout.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Logout = props =>{
+    return (
+        <button onClick={props.onClick} className={props.className}>
+            Logout
+        </button>
+    )
+}
+
+export default Logout;
